requirements.txt
Flask==1.0.2
connexion[swagger-ui]==2.2.0
SQLAlchemy==1.3.3
psycopg2==2.8.2
pyyaml==5.1
PyJWT==1.7.1
cryptography==2.7